infra/requirements.txt
aws-cdk-lib==2.173.2
aws-cdk.aws-amplify-alpha==2.177.0a0
constructs>=10.0.0,<11.0.0
